+++ settings.gradle
@@ -1,1 +1,1 @@
-rootProject.name = 'delievery'
+rootProject.name = 'delivery'
