+++ src/main/java/com/dia/delivery/store/dto/StoreOneResponseDto.java
@@ -1,18 +1,23 @@
-package com.dia.delievery.store.dto;
+package com.dia.delivery.store.dto;
 
-import com.dia.delievery.store.entity.Stores;
+import com.dia.delivery.review.entity.Reviews;
+import com.dia.delivery.store.entity.Stores;
 import lombok.Getter;
 
+import java.util.List;
+
 @Getter
-public class StoreResponseDto {
+public class StoreOneResponseDto {
     private String name;
     private String introduction;
     private String imageUrl;
     private String category;
-    public StoreResponseDto(Stores store) {
+    private List<Reviews> reviewsList;
+    public StoreOneResponseDto(Stores store) {
         this.name = store.getName();
         this.introduction = store.getIntroduction();
         this.imageUrl = store.getImageUrl();
         this.category = store.getCategory();
+        this.reviewsList = store.getReviewsList();
     }
 }
