+++ src/test/java/com/dia/delivery/DelieveryApplicationTests.java
@@ -4,7 +4,7 @@ import org.junit.jupiter.api.Test;
 import org.springframework.boot.test.context.SpringBootTest;
 
 @SpringBootTest
-class DelieveryApplicationTests {
+class DeliveryApplicationTests {
 
     @Test
     void contextLoads() {
